u128 -> u64, forgot dbs dont support it LMAO

+++ src/models/transactions_model.rs
@@ -1,4 +1,4 @@
-use anyhow::{Context, Result};
+use anyhow::Result;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
@@ -40,16 +40,18 @@ pub enum TransactionData {
 pub struct SwapData {
     pub token_in: String,
     pub token_out: String,
-    pub amount_in: u128,
-    pub amount_out: u128,
+    // DB only supports up to 2^64
+    pub amount_in: u64,
+    pub amount_out: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct LiquidityData {
     pub token_a: String,
     pub token_b: String,
-    pub amount_a: u128,
-    pub amount_b: u128,
+    // DB only supports up to 2^64
+    pub amount_a: u64,
+    pub amount_b: u64,
     pub liquidity_amount: String,
     pub tick_lower: Option<i32>,
     pub tick_upper: Option<i32>,
